+++ src/main/java/com/algoanalyzer/domain/problem/dto/request/ProblemRequestDto.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class ProblemRequestDto {
     @NotNull(message = "문제 번호는 필수입니다.")
     @Positive(message = "문제 번호는 양수여야 합니다.")
-    private Long problemId;
+    private int problemId;
 }
